+++ src/mn_menu.h
@@ -90,6 +90,7 @@ void MN_DisableVoxelsRenderingItem(void);
 void MN_UpdateDynamicResolutionItem(void);
 void MN_DisableResolutionScaleItem(void);
 void MN_UpdateFpsLimitItem(void);
+void MN_UpdateCSStatsTrackingItem(void); // [Cherry]
 
 typedef enum
 {